+++ include/qristal/core/python/py_noise_model.hpp
@@ -2,6 +2,10 @@
 
 #pragma once
 #include <pybind11/pybind11.h>
+#include "qristal/core/noise_model/readout_error.hpp"
+
+PYBIND11_MAKE_OPAQUE(qristal::ReadoutError);
+PYBIND11_MAKE_OPAQUE(std::unordered_map<size_t, qristal::ReadoutError>);
 
 namespace qristal {
   /// Bind noise modelling-related types